+++ source/component.d.ts
@@ -33,6 +33,34 @@ export declare class Component extends HTMLElement {
 	 */
 	dispatch(action: Record<string, any>): void;
 
+	/**
+	 * Loads the value from the given attribute and tries to convert it to an array.
+	 * @param attributeName The attribute name where the data should be taken from.
+	 * @returns An array if the conversion was successful. If the value can't be converted or the attribute does not exist, null.
+	 */
+	getArrayAttribute(attributeName: string): Array | null;
+
+	/**
+	 * Loads the value from the given attribute and tries to convert it to a boolean value.
+	 * @param attributeName The attribute name where the data should be taken from.
+	 * @returns true if the value is "true", false if the value is "false", or null if the value cannot be converted.
+	 */
+	getBooleanAttribute(attributeName: string): boolean | null;
+
+	/**
+	 * Loads the value from the given attribute and tries to convert it to a number.
+	 * @param attributeName The attribute name where the data should be taken from.
+	 * @returns A number if the conversion was successful, or null if the value can't be converted.
+	 */
+	getNumberAttribute(attributeName: string): number | null;
+
+	/**
+	 * Loads the value from the given attribute and tries to convert it to a JavaScript object.
+	 * @param attributeName The attribute name where the data should be taken from.
+	 * @returns A JavaScript object if the conversion was successful, or null if the value can't be converted.
+	 */
+	getObjectAttribute(attributeName: string): Record<string, any> | null;
+
 	// Properties (using Proxy)
 	dependencies: {
 		store?: {