+++ source/component.js
@@ -35,7 +35,7 @@ export class Component extends HTMLElement {
 		componentClass.prototype.dependencies = new Proxy({}, {
 			get(dependencyContainer, propertyName, ...args) {
 				if (!(propertyName in dependencyContainer) && propertyName !== "store") {
-					console.warn(`${tagName} is accessing the undefined dependency '${propertyName}'. Did you inject it using Injector.injectDependencies()?`);
+					console.warn(`${tagName}: Tried to access dependency '${propertyName}, which is not injected'`);
 				}
 
 				return Reflect.get(dependencyContainer, propertyName, ...args);
@@ -66,10 +66,46 @@ export class Component extends HTMLElement {
 	 */
 	dispatch(action) {
 		if (this.dependencies.store === undefined) {
-			throw new Error(`No store was injected into ${this.tagName.toLowerCase()}, so the action ${action?.type} could not be dispatched. Did you inject the store using Injector.injectDependencies()?`);
+			throw new Error(`${this.tagName.toLowerCase()}: No store was injected, so the action ${action?.type} could not be dispatched`);
 		}
 
 		this.dependencies.store.dispatch(action);
+	}
+
+	/**
+	 * Loads the value from the given attribute and tries to convert it to an array.
+	 * @param {string} attributeName The attribute name where the data should be taken from.
+	 * @returns {?Object[]} An array if the conversion was successful. If the value can't be converted or the attribute does not exist, null.
+	 */
+	getArrayAttribute(attributeName) {
+		return this.getTypedAttribute(attributeName, "array");
+	}
+
+	/**
+	 * Loads the value from the given attribute and tries to convert it to a boolean value.
+	 * @param {string} attributeName The attribute name where the data should be taken from.
+	 * @returns {?boolean} true, when the value is "true", false when the value is "false". If the value can't be converted or the attribute does not exist, null.
+	 */
+	getBooleanAttribute(attributeName) {
+		return this.getTypedAttribute(attributeName, "boolean");
+	}
+
+	/**
+	 * Loads the value from the given attribute and tries to convert it to a number.
+	 * @param attributeName The attribute name where the data should be taken from.
+	 * @returns A number if the conversion was successful, or null if the value can't be converted.
+	 */
+	getNumberAttribute(attributeName) {
+		return this.getTypedAttribute(attributeName, "number");
+	}
+
+	/**
+	 * Loads the value from the given attribute and tries to convert it to a JavaScript object.
+	 * @param {string} attributeName The attribute name where the data should be taken from.
+	 * @returns {?Object} A JavaScript object if the conversion was successful. If the value can't be converted or the attribute does not exist, null.
+	 */
+	getObjectAttribute(attributeName) {
+		return this.getTypedAttribute(attributeName, "object");
 	}
 
 	/**
@@ -88,10 +124,10 @@ export class Component extends HTMLElement {
 			if (parsedValue?.constructor.name.toLowerCase() === destinationType) {
 				return parsedValue;
 			}
-		} catch {}
+		} catch { /*eslint no-empty: "off"*/ }
 
 		return null;
-	  }
+	}
 
 	/**
 	* Connects all attributes starting with "$" to functions of the parent component.
@@ -113,7 +149,7 @@ export class Component extends HTMLElement {
 					childNode.addEventListener(eventName.substring(1), (event) => component[functionName](event));
 				}
 				else {
-					console.warn(`${component.tagName.toLowerCase()} defined a ${eventName} handler called '${functionName}' which does not exist or isn't a function`);
+					console.warn(`${component.tagName.toLowerCase()}: ${eventName} handler '${functionName}' doesn't exist or is not a function`);
 				}
 			}
 		});